chore: support define local variables (#115)
Submit a pull request for this project.

<!-- If you have an Issue that related to this Pull Request, you can
copy this Issue's description -->

# Why? 
<!-- 
> Related to which issue?
> Why we need this pull request?
> What is the user story for this pull request? 
-->
Managing local configuration and environment variables can be
challenging, especially when everyone's setup is different. It can be
tempting to modify the .env.defaults file to suit our needs, but this
can cause problems if we forget to restore it before pushing our
changes. It's important to find a convenient and consistent method for
managing these variables to avoid any misoperation or unintended
changes.

# What?
<!-- 
> Can you describe this feature in detail?
> Who can benefit from it? 
-->
Create a `.env.development.local` file in the `env` folder to define
environment variables for local development, using local variables can
help reduce the risk of misoperation, and we can use it to store
sensitive or environment-specific information that you don't want to
commit to your repository

# How?
<!-- 
> Do you have a simple description of how this pull request is
implemented?
-->
- [x] move `/env` folder to `/root` folder
- [x] add the `.env.development.local` file for local development
- [x] add the content of "how to define local variables" in `readme`