chore: optimise email verify html (#184)
# Why? 
Improve the design and layout of the email verify HTML to make it more
visually appealing and user-friendly. The current design is too basic
and only contains the text.

# What?
Make it more interactive and engaging, with appropriate use of images
and other multimedia elements if relevant. Additionally, consider adding
helpful prompts and instructions to guide users through the verification
process.

# How?
Add an image and replace the old text.