fix: cdn anti-leech in the development environment (#5735)
The Referrer Policy is a security feature that determines how much
information about the referring URL (referrer) is included when making a
request from one webpage to another. It controls whether the referrer is
sent or not, and if sent, what information is included.

In HTML, there are several ways to modify the Referrer Policy:

Using the meta element: You can add a meta element inside the head tag
of your HTML file. The meta element should have the name attribute set
to "referrer" and the content attribute set to the desired Referrer
Policy value. For example:
```html
<head>
  <meta name="referrer" content="no-referrer">
</head>
```
This sets the Referrer Policy to "no-referrer" which means no referrer
information will be sent.

Using the Referrer-Policy HTTP header: You can set the Referrer Policy
using the Referrer-Policy HTTP header in the server response. The header
value should be set to the desired Referrer Policy. For example:
```yaml
Referrer-Policy: no-referrer
```
This sets the Referrer Policy to "no-referrer".

Using the Content-Security-Policy HTTP header: The
Content-Security-Policy HTTP header can also be used to set the Referrer
Policy. This header provides a way to set multiple security-related
policies, including the Referrer Policy. For example:
```css
Content-Security-Policy: referrer no-referrer;
```
This sets the Referrer Policy to "no-referrer".

It's important to note that the specific Referrer Policy values and
their behavior may vary across different browsers. Therefore, it's
recommended to refer to the browser documentation to understand the
supported values and their implications.

Co-authored-by: Kilian <wangkailang@users.noreply.github.com>