Feature/fix automation filter readonly (#7040)
close https://github.com/vikadata/vikadata/issues/7032
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: The `FilterCascader` component now supports a new `disabled`
prop, preventing interaction when disabled.
- Bug Fix: The `FilterDateDuration` component in `FilterDate` can now be
disabled based on the `isViewLock` variable.
- New Feature: The `FilterDateDuration` component introduces an optional
`disabled` prop to disable the component if provided and set to `true`.
- Refactor: The `FilterGeneralSelect` component has undergone changes in
import statements, rendering, and event handling, improving the UI and
user experience.
- Refactor: The `FilterOptions` component now receives the `disabled`
prop instead of `isViewLock`, affecting the behavior of filter options
when disabled.
- Bug Fix: The `FilterValue` component now supports the `disabled` prop
for `FilterOptions` and `FilterCascader`, disabling certain filter
options when the component is locked or disabled.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->