chore(backend-server): migration resources ro parameters deserialize (#7060)
<!-- 
# What?
 Migrate attachment resource interfaces and add custom deserialization
-->


<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added a custom deserializer class
`StringOrArrayToStringArrayDeserializer` to handle JSON values that can
be either a string or an array of strings. This allows for more flexible
parsing of JSON data.
- New Feature: Updated the `MigrationResourcesRo` class to use the
custom deserializer for the `resourceKeys` field, enabling conversion of
a string or an array of strings to a list of strings during JSON
deserialization. This improves the handling of resource keys in the
application.

<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->