Feat/chat bot (#7196)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added new string keys related to AI credit time
dimensions in the `stringkeys.interface.ts` file.
- Style: Modified the styling and rendering of the chart displayed by
the `CreditCostCard` component. This includes adjustments to margins,
ticks, and dots of the X and Y axes, as well as changes to stroke color
and width of lines in the chart.
- Refactor: Introduced changes to import statements and labels of
options in the `SELECT_LIST` array in the `enum.ts` file. Added an
import statement for `@apitable/core` and replaced labels with
translated strings from the `Strings` module.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->