fix(backend-server): repair preview file prompt does not exist (#7162)
<!-- 
# What?
repair preview file prompt does not exist 
-->


<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Fixed an issue where the preview file prompt was not
appearing. The changes in `OfficeServiceImpl.java` handle different
cases for constructing the `originUrl` and properly encode it to prevent
conversion failures for special filenames, improving the reliability of
generating the file URL for previewing.
- Refactor: Modified the `spliceAssetUrl` method in
`ConstProperties.java` to use the `resourceUrl` property of
`getOssBucketByAsset()` instead of `this.getOssBucketByAsset()`. This
change may affect the behavior of the code.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->