Feat/chat bot (#7225)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added support for the `@apitable/core` module in the
`init.ts` file of the `shared/player` module.
- New Feature: Modified the `chat_completion` object in the Python
example code to include properties such as `conversation id`, `actions`,
and `model`.
- New Feature: Enhanced the `Modal` component in `modal.tsx` to
conditionally render a `PreviewMedia` component or an iframe based on
the value of `isYoutebeVideo`.
- Style: Updated the typography variant of a component in
`usage_warn_modal.tsx` from 'h5' to 'body1', affecting the styling of
the alert content in the UsageWarnModalInner component.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->