Feat/chat bot (#7226)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added a new string key
`ai_feedback_insight_pagination_total` to the `StringKeysMapType` type
in the core configuration. This enhances the flexibility of language
localization.
- Refactor: Updated ESLint configuration in the datasheet package to
improve code quality and enforce best practices for imports, comma
spacing, restricted imports, and unused variables in TypeScript.
- Refactor: Imported `@apitable/core` and added a TypeScript ignore
comment in the `init.ts` file of the shared player module. No changes to
functionality or behavior.
- Documentation: Updated the Python example in the chat tool bar API
documentation to include additional properties (`conversation id`,
`actions`, `model`, `total cost`, and `result`) for the
`chat_completion` object.
- Refactor: Modified the `Modal` component in the UI guide to
conditionally render a `PreviewMedia` component or an iframe based on
the value of `isYoutebeVideo`. Added an `//@ts-ignore` comment to
suppress TypeScript errors.
- Style: Changed the variant of a Typography component from 'h5' to
'body1' in the UsageWarnModalInner component, affecting the styling of
the alert content.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->