fix: remove stylelint (#7551)
# Human says

Remove stylelint
Related: https://github.com/vikadata/vikadata/issues/7548

# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Style: Removed "stylelint.vscode-stylelint" extension from the list of
installed extensions in the VS Code configuration.
- Style: Removed configuration for `stylelint` from the VS Code
settings.
- Refactor: Removed the stylelint plugin "plugin/font-weight-no-number"
that checked for numeric values in the `font-weight` property.
- Style: Removed stylelint-disable comments from CSS files.

These changes primarily focus on removing stylelint-related
configurations and plugins. They do not introduce new features or alter
the functionality of the code. However, it's important to note that the
removal of the stylelint extension and plugin may impact linting and
styling checks for CSS files.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->