Fix/freeze column optimization (#7571)
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added a new key-value pair `'freeze_column_reset'` to the
`StringKeysMapType` type in the `core` package configuration.
- New Feature: Imported the `FieldIcon` component in the
`field_head.tsx` file of the `datasheet` package.
- Bug Fix: Modified the `onFrozenColumn` function in the `dom_grid.tsx`
file of the `datasheet` package to accept an additional parameter
`reset`, which defaults to `false`. If `reset` is `true`, the
`columnIndex` is set to 0.
- Style: Replaced the import statement for `blackBlue` with `ThemeName`
in the `use_grid.tsx` file of the `datasheet` package. Adjusted the
`shadowProps` object and shadow properties based on the `themeName`
variable.
- Refactor: Modified the signature of the `onFrozenColumn` function in
the `context_menu.tsx` file of the `multi_grid` component. Added a new
parameter `reset` of type boolean to the function.
- Refactor: Updated the `FieldMenu` component in the `field_menu.tsx`
file of the `multi_grid` component. Added a new parameter `reset` to the
`onFrozenColumn` callback function and adjusted its usage. The freeze
column menu item now displays "Reset" when the current column is the
last frozen column.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->