fix(backend-server): imposes restrictions on inviting members (#7714)
<!-- 
# What?
The space of the domestic version of vika is strictly limited to seats,
and excessive invitations are not allowed; Foreign version of apibale
can be overinvited
-->

<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added a new configuration property
`BILLING_APITABLE_ENABLED` to the `SpaceServiceImpl` class. This
property allows skipping seat over-limit validation for paid spaces in
apitable mode. If set to `true`, the validation is skipped; otherwise,
it is performed as usual. This enhancement provides more flexibility for
managing seat limits in different scenarios.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->