fix: l10n and crash issues (#7832)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Improved error handling in the
`getDisplayResultWithComponent` function of the React component in
`enterprise/datasheet/chat/cui/index.tsx`. Added null and undefined
checks for the `data` parameter to prevent potential issues. If `data`
is null or undefined, an empty string is returned. This ensures that the
function handles null or undefined values correctly, improving the
reliability of the component.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->