feat: ai training history (#7828)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added AI training history translation keys for displaying
descriptions and column titles in the AI training history table.
- Bug Fix: Updated the URL in the `submitQuestionnaire` function to
`'https://workfun.aitable.ai/feedback-form'`.
- Style: Added a space after the closing curly brace in the
`deleteRoleMember` function.
- Documentation: Modified the `beforeCapture` prop of the
`Sentry.ErrorBoundary` component, changing the type from `unknown` to
`any`.
- Refactor: Refactored the `TaskItem` React component in the automation
run history list, including changes to rendering the status icon,
conditional rendering for the "More" button, and styling updates.
- Style: Modified CSS styles of the pagination component, including
color, background color, opacity, and button styling.
- Refactor: Modified the condition in the `getAvInfoRequestUrl` function
to check if the `search` property exists on the URL object instead of
checking the size of the search parameters.
- New Feature: Added a new method `retrieveTrainings` to the
`AiController` class for retrieving a list of training information for a
specific AI.
- New Feature: Added a new class `TrainingInfoList` representing a list
of training information.
- New Feature: Introduced a new interface `IAiTrainingService` with a
`getTrainingList` method for retrieving a list of `TrainingInfoVO`
objects.
- New Feature: Added an implementation of the `getTrainingList` method
in the `AiTrainingServiceImpl` class, retrieving training information
using an external API and converting it into a list of `TrainingInfoVO`
objects.
- Chore: Added an import statement for `IAiTrainingService` and injected
it as a dependency in the test class.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->

---------

Co-authored-by: Shawn Deng <dengguiheng@apitable.com>
Co-authored-by: liuyi <liuyi@vikadata.com>
Co-authored-by: Chambers <chenbochao@vikadata.com>
Co-authored-by: Kelly Peilin Chan <kelly@apitable.com>
Co-authored-by: Harry Yep <git@hode.co.uk>