fix: some error (#7856)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Added an `else` block in the `use_invite.tsx` file to handle
a specific error condition when the API response is not successful. If
the error code is `OVER_LIMIT`, a usage alert is triggered.
- Refactor: Renamed the property from `tokens` to `words` in the
returned object of the `History` component in the `index.tsx` file. This
change affects the structure of the data being rendered.
- Refactor: Modified the `dataIndex` property and increased the width of
the second column in the `columns` array in the `config.ts` file of the
`history_detail_table` component. The property was changed from
`'tokens'` to `'words'`, and the width was increased from 50 to 70.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->

---------

Co-authored-by: Kelvin Poon <kelvin@apitable.com>