refactor: entering the space station no longer directly requests memb… (#8022)
…er data


# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Introduced a custom hook `useGetMemberStash` to fetch
member stash data for a specific space. This hook utilizes React's
`useEffect` and `useState` hooks, as well as Redux's `useSelector` hook.
It improves the loading state management and returns the fetched data.
- Refactor: Added a new method `hasCacheId()` to the `MemberStash`
class, which checks if the current cache ID matches the provided space
ID. This method enhances the ability to determine if the cache ID is
already set for a particular space.
- Refactor: Made changes to import order in `space_id.ts`, specifically
for `CacheManager` and `StoreActions`. Removed an import statement for
`memberStash` from a specific module. Commented out the line
`memberStash.loadMemberList(spaceId)`.
- Refactor: Modified the `ICommonListProps` interface by removing the
`keyword` parameter from the `onSearchChange` function and adding a new
optional `isLoadingData` boolean property.
- Refactor: Updated the `CommonList` component to conditionally render a
loading indicator when `isLoadingData` is true. Improved logic for
rendering "No Data" or "No Search Result" messages.
- Refactor: Modified the `MemberOptionList` component by introducing the
`useGetMemberStash` hook to fetch member stash data. Improved handling
of loading state with the `memberLoading` state variable. Updated the
`initList` variable to use `memberStashList` from the hook. Passed
`isLoadingData` prop to the `AutoComplete` component.
- Refactor: Replaced the previous implementation in `filter_member.tsx`
with the `useGetMemberStash` hook to retrieve the `memberStashList`
value. Changed the import statement for `memberStash` to the new custom
hook import.
- Refactor: Updated the `date_time_alarm.tsx` file by replacing the
previous implementation with the `useGetMemberStash` hook to retrieve
the `memberStashList` value. Changed the import statement for the hook.
- Refactor: Modified the `mobile_alarm.tsx` file by replacing the
previous implementation of `stashList` with a call to the
`useGetMemberStash` hook, which returns the `memberStashList`. Changed
the import statement for the hook.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->