Feat/archive row (#8131)
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added two new entries, `'ArchiveRecords'` and
`'UnarchiveRecord'`, to the `StringKeysMapType` type in the `core`
package's `stringkeys.interface.ts` file. This may impact the external
interface or behavior of the code.
- Bug Fix: Modified the `ArchivedRecords` component in the `datasheet`
package's `index.tsx` file. Two cases in the switch statement now return
a JSON string representation of the `cellValue` instead of the original
value. This ensures consistent output for archived records.
- Style: Made a whitespace change by adding an extra line in the
`ArchivedRecords` component in the `datasheet` package's `index.tsx`
file.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->

---------

Co-authored-by: Harry Yep <git@hode.co.uk>