fix: archived action button ui (#8175)
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Style: Modified the alignment and hover behavior of elements, added a
margin to a span element, changed the background color and border color
of table cells, and adjusted the opacity of a loading spinner in the
`archive_record` component's CSS styles.
- New Feature: Made changes to the `archive_record` component in the
`apitable` package. Updated import statements, implemented table
pagination, added ellipsis to table cells, and modified tooltips and
modals for an improved user experience. Increased the page size for the
table.

<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->