Refactor/welcome (#8220)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Fixed an issue in the `RobotListItemCard` component where the
`onClick` event handler was not passing the correct value to the
`toggleRobotActive` function, resulting in incorrect behavior. The fix
ensures that the correct value is now passed and the updated state is
reflected correctly.
- Refactor: Modified the conditional rendering in the `Welcome`
component to render either the `ChatWelcome` or `Guide` component based
on the value of `getEnvVariables().IS_AITABLE`. This change may affect
the behavior of the code depending on the value returned by
`getEnvVariables().IS_AITABLE`.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->

---------

Co-authored-by: Caedman Ziwen Lan <lanziwen@apitable.com>