Hotfix/cola (#8229)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Style: Added a new class name to the `<Loading>` component in the
`DataSourceSelectorForNode` module, which applies a backdrop blur effect
on the loading overlay when `isLoadingExtraData` is true.
- Style: Corrected the indentation of a conditional statement in the
`HiddenField` component, aligning it with the previous line. This change
does not affect the logic or functionality of the code.
- Bug Fix: Modified the `src` attribute of an `<iframe>` element in the
`ChatWelcome` component to include a query parameter
`aitable_widget=chat`. This may impact the behavior or appearance of the
embedded content within the iframe.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->