fix: calendar hightlight style (#8241)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Refactor: Modified the `onGlobalMouseDown` function in the
`CalendarView` component to only clear the selection and active row info
if there is an active cell. This improves the behavior of the calendar
view.
- Style: Updated the CSS styles for calendar highlighting, replacing the
background color with a linear gradient and updating the border color.
This enhances the visual appearance of the calendar view.
- Refactor: Removed the unused import statement and the
`isClickDragDropModal` function from the codebase. The function was
responsible for determining whether a click event occurred within
specific elements on the page but was no longer needed.
- Chore: Cleaned up the codebase by removing unnecessary code and
improving code organization.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->