fix: web-server ci (#8391)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Chore: Updated exclusion pattern in `.gitignore` to reflect changes in
file paths.
- Chore: Modified `Makefile` to sync and copy the `ai-components`
directory instead of `ai` in the `enterprise-rsync` and
`enterprise-copy` targets. Added a line to append "IS_ENTERPRISE=true"
to `.env.local` in the `apitable/packages/datasheet` directory.
- Chore: Corrected the location of the `package.json` file for the
`ai-components` package in `Dockerfile.web-server`.

These changes are primarily maintenance-related and do not introduce new
features or affect the functionality of the code.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->