fix: workdoc image selected (#8518)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Improved accuracy of filter conditions in the `WorkdocField`
class by updating the logic for the `Contains` and `DoesNotContain`
operators.
- Style: Introduced a new import statement for the `classNames` library
and added a conditional class name to the `NodeViewWrapper` component
based on the `selected` prop. This allows for applying a CSS class when
the component is selected, enabling easier styling customization.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->