fix: workdoc expand large text (#8526)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Disabled a shortcut when editing a specific type of field
(`FieldType.Workdoc`) to ensure the shortcut is not active when editing
a workdoc field.
- Style: Added a `margin-top` property to align an SVG element in the
CSS styling of a component.
- Style: Modified the class name and added properties to the CSS styles
of the selected image node in a ProseMirror editor, potentially
affecting the styling or behavior of the component.
- Style: Added the `flex: none;` property to the `.right` class in the
CSS file to prevent the element from growing or shrinking when the
container is resized, impacting the alignment of elements within the
container.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->